Changing things around and learning python
