--- lambda/py/languageStrings.py
@@ -1,80 +1,0 @@
-languageStrings : {
-    'en' : {
-        'translation' : {
-            'SKILL_NAME' : 'Overwatch League',
-            'WELCOME_MSG' : 'Welcome to Overwatch League. You can ask me about Overwatch League games and standings. For example,\
-                you can ask when the next %s game is, or, who the top three teams are. You can also ask me what games are today, or,\
-                what yesterday\'s results were. I am also able to tell you the games that will be played tomorrow. For more help,\
-                say help.',
-
-            'WELCOME_REPROMPT' : 'For match information say, when is the next %s game, or, for ranking information ask what are\
-                the league standings. For more help, say help.',
-
-            'PERMISSIONS_WELCOME' : 'Welcome to Overwatch League.',
-
-            'PERMISSIONS_PROMPT': 'In order to get match times in your local time I need permission to access your device information and location.\
-                Use the Alexa companion app and enable permissions by either selecting the manage permissions link on the most recent card,\
-                or, you can change permissions under this skills settings. Afterwards, please try your request again.',
-
-            'STANDINGS_REPROMPT' : 'How many teams did you want to know about?',
-            'INVALID_TEAM_MSG' : 'I am sorry, %s is not a valid league team. Can I Help by searching for another team?',
-            'INVALID_STAND_MSG': 'I am sorry, I did\'t quiet get that.',
-            'TEAM_REPROMPT' : 'Can I help by searching for another team?',
-            'API_ERROR_MSG' : 'Something went wrong with an API, please try again later.',
-            'OWL_API_ERR_MSG' : 'It looks like information from Overwatch League is unavailable right now, please try again later.',
-            'GOOG_API_ERR_MSG' : 'It looks like I am having a hard time getting match times in your local timezone. Please check\
-                that you have allowed Overwatch League to use your location in the Alexa companion app and try again.',
-
-            'HELP_MSG' : 'To get match date and times for a specific team, you could ask, when is the next %s game. To get\
-                the next game regardless of the team, you could say tell me the next match, or, when is the next game. For\
-                standings information, you can ask, who are the top teams, and I will tell you the current top three teams.\
-                You may also ask for any number of teams like the top four or six teams. I am also able to tell you the current\
-                record for any team. If you would like to know the current standings for every team, you can ask for all standings.\
-                I am also able to tell you today\'s games, yesterday\'s results, and, tomorrow\'s games.',
-
-                'HELP_REPROMPT' : 'Say help to hear this again.',
-
-            'SIMPLE_REPROMPT' : "Is there something else you would like to know?",
-            'SHUTDOWN_MSG' : 'The world could always use more heros.'
-        }
-    },
-    'en-US' : {
-        'translation' : {
-            'SKILL_NAME' : 'Overwatch League',
-            'WELCOME_MSG' : 'Welcome to Overwatch League. You can ask me about Overwatch League games and standings. For example,\
-                you can ask when the next %s game is, or, who the top three teams are. You can also ask me what games are today, or,\
-                what yesterday\'s results were. I am also able to tell you the games that will be played tomorrow. For more help,\
-                say help.',
-
-            'WELCOME_REPROMPT' : 'For match information say, when is the next %s game, or, for ranking information ask what are\
-                the league standings. For more help, say help.',
-
-            'PERMISSIONS_WELCOME' : 'Welcome to Overwatch League.',
-
-            'PERMISSIONS_PROMPT': 'In order to get match times in your local time I need permission to access your device information and location.\
-                Use the Alexa companion app and enable permissions by either selecting the manage permissions link on the most recent card,\
-                or, you can change permissions under this skills settings. Afterwards, please try your request again.',
-
-            'STANDINGS_REPROMPT' : 'How many teams did you want to know about?',
-            'INVALID_TEAM_MSG' : 'I am sorry, %s is not a valid league team. Can I Help by searching for another team?',
-            'INVALID_STAND_MSG': 'I am sorry, I did\'t quiet get that.',
-            'TEAM_REPROMPT' : 'Can I help by searching for another team?',
-            'API_ERROR_MSG' : 'Something went wrong with an API, please try again later.',
-            'OWL_API_ERR_MSG' : 'It looks like information from Overwatch League is unavailable right now, please try again later.',
-            'GOOG_API_ERR_MSG' : 'It looks like I am having a hard time getting match times in your local timezone. Please check\
-                that you have allowed Overwatch League to use your location in the Alexa companion app and try again.',
-
-            'HELP_MSG' : 'To get match date and times for a specific team, you could ask, when is the next %s game. To get\
-                the next game regardless of the team, you could say tell me the next match, or, when is the next game. For\
-                standings information, you can ask, who are the top teams, and I will tell you the current top three teams.\
-                You may also ask for any number of teams like the top four or six teams. I am also able to tell you the current\
-                record for any team. If you would like to know the current standings for every team, you can ask for all standings.\
-                I am also able to tell you today\'s games, yesterday\'s results, and, tomorrow\'s games.',
-
-                'HELP_REPROMPT' : 'Say help to hear this again.',
-
-            'SIMPLE_REPROMPT' : "Is there something else you would like to know?",
-            'SHUTDOWN_MSG' : 'The world could always use more heros.'
-        }
-    }	 
-}
